requirements.txt
pygame>=2.5.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
